Add data indirection for bytestrings in synthetic
ProxyStore and Ray do not place nicely together with Proxy[bytes] types.
This is because Ray skips serializing bytes types, but then crashes
because the Proxy isn't actually an instance of bytes. To get around
this, we just wrap bytestring data in the synthetic app in another class
so that the Proxy[Data] gets correctly serialized.