Ignore extra options in engine plugin configs
Forbidding extra options in the engine plugin configs (executor, filter,
and transformer) meant that trying to override the plugin type defined
in a config file from a CLI argument would raise an "extra options
forbidden" error by the pydantic validators because of the options for
the not used plugin variant in the config.

This change should support that use case. Note that the run Config and
AppConfig still forbid extra options as a sanity check, and those don't
have the same issue of wanting to frequently switch the plugin type.