Fix pickle task name clash (#121)
Decorating a top-level function with @task would replace the function
with a Task that references the function. However, the Task would be
updated with functools.update_wrapper to look like the function. This
included the __name__ and __module__ which essentially clobbered the
original function so pickling Task would fail because the reference to
the wrapped function did not resolve to the right thing anymore.

This was fixed by switching to decorating with functions only (via
functools' wraps and partial). I.e., there is not Task object now (well
there is but it's just a protocol for what the wrapper functions do).
There's some tricky typing and serialization consequences that I tried
to leave as many comments for and added serializatin and typing tests.