Remove all filter/null transformer (#123)
Previously, the filter defaulted to an AllFilter and the transformer to
a NullTransformer. This was nice from the perspective of code
cleanliness because there were no branches for "is there a filter" or
"is there a transformer." However, this did technically mean that
everything was transformed (the transformer was an identity function).

So when addressing #123 in a prior commit, this meant that using the
default config for TaPS, every argument was getting "transformed" by the
NullTransformer, resulting in tons of useless log lines.

This commit changes the default filter and transformer to be the literal
value None, rather than no-op implementations. The corresponding
arguments to Engine, EngineConfig, and TaskTransformer were updated to
support None values. Fortunately, the TaskTransformer wrapper means that
all logic changes were isolated there (i.e., the branches for if a
filter/transformer were configured).

A tricky side-effect of this change is that _make_config_cls expects all
engine plugins to have a value so this required some changes to support
the None value path. Got that all fixed up and along the way improved
some of the outputs in the CLI help.