Forbid extra fields in engine plugin config (#124)
This is a revert of df1be9fe6b6864645a7285a892578d0ad2f19927. The
justification there to ignore extra fields is overriding the plugin kind
in a config file from the CLI args. This is a nice feature, but it comes
at the risk of not catching typos in config field names. I think it's
ultimately better to be overly strict at the cost of some flexibility
over letting simple and subtle errors in a config sneak in.